HAN-140 feat evaluate api
